backend/requirements.txt

Flask==2.3.2
Flask-Cors==4.0.0
Werkzeug==2.3.6
Pillow==9.5.0
torch==2.6.0
transformers==4.50.3
requests==2.31.0
diffusers==0.32.2
accelerate==1.6.0
huggingface-hub>=0.26.0,<1.0
pymongo==4.6.1
cloudinary==1.36.0
python-dotenv==1.0.1
